Added threading
Added a bunch of threading stuff. I don't know if it works, though,
because I haven't tried connecting to it yet. I updated the CMakeLists
file to link against the correct stuff, although I don't know if it's
portable. I also had a mutex system planned for some stuff but never
added it.